requirements.txt
numpy~=1.19.1
matplotlib~=3.3.1
scikit-image~=0.17.2
torchsummary~=1.5.1